Minor files restructuring...
